// frontend/lib/generated/InvestigationCampaignPageResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InvestigationCampaignListItemView } from "./InvestigationCampaignListItemView";
import type { InvestigationProjectionEnvelope } from "./InvestigationProjectionEnvelope";

export type InvestigationCampaignPageResponse = { envelope: InvestigationProjectionEnvelope, campaigns: Array<InvestigationCampaignListItemView>, };