// frontend/lib/generated/InvestigationHypothesisListView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InvestigationHypothesisListItemView } from "./InvestigationHypothesisListItemView";
import type { InvestigationProjectionEnvelope } from "./InvestigationProjectionEnvelope";

export type InvestigationHypothesisListView = { envelope: InvestigationProjectionEnvelope, hypotheses: Array<InvestigationHypothesisListItemView>, };