// frontend/lib/generated/InvestigationRequestStopRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type InvestigationRequestStopRequest = { sessionId: string, operationId: string, stageExecutionId: string, stageRunRequestId: string, expectedInvestigationRunStateHead: string, expectedChangeSeq: number, idempotencyKey: string, };