// frontend/lib/generated/InvestigationTimelinePageResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InvestigationProjectionEnvelope } from "./InvestigationProjectionEnvelope";
import type { InvestigationTimelineItemView } from "./InvestigationTimelineItemView";

export type InvestigationTimelinePageResponse = { envelope: InvestigationProjectionEnvelope, events: Array<InvestigationTimelineItemView>, };